Updated CollapsingToolbar Animation Added Recycler with Actors

+++ app/src/main/java/com/thebaileybrew/flix/model/Movie.java
@@ -87,7 +87,7 @@ public class Movie implements Parcelable{
         return movieVoteCount;
     }
 
-    public float getMovieVoteAverage() {
+    public long getMovieVoteAverage() {
         return movieVoteAverage;
     }
